Refactor: Updated Modal Components Folder Structure
Organize modals in dashboardV2 by feature area to improve code organization and maintainability.

- Move project-related modals to modals/project/:
  - CreateProjectModal
  - UpdateProjectModal
  - DeleteProjectModal
  - ProjectName
  - ProjectDescription
  - VisibilitySelector

- Move task-related modals to modals/task/:
  - CreateTaskModal

- Update import paths in:
  - ProjectCardV2.tsx
  - ProjectsOverviewV2.tsx
  - TasksOverview.tsx

All internal imports within each subdirectory remain unchanged as they use relative paths.
